Send a notification with the linting result 📳
