+++ src/notifier.mjs
@@ -1,0 +1,36 @@
+import notifier from 'node-notifier'
+
+import { pluralise } from './utils.mjs'
+
+const notifyResults = results => {
+  let totalErrorCount = results.reduce((total, { processedResult: { errorCount = 0 } }) => total + errorCount, 0)
+  let totalWarningCount = results.reduce((total, { processedResult: { warningCount = 0 } }) => total + warningCount, 0)
+
+  if (totalErrorCount > 0) {
+    notifier.notify({
+      message: `${totalErrorCount} ${pluralise('error', totalErrorCount)} found. Please fix ${totalErrorCount > 1 ? 'them ' : 'it '}before continuing.`,
+      sound: 'Frog',
+      title: '🚨 Lint Pilot 🚨',
+    })
+    return
+  }
+
+  if (totalWarningCount > 0) {
+    notifier.notify({
+      message: `${totalWarningCount} ${pluralise('warning', totalWarningCount)} found. Please review ${totalWarningCount > 1 ? 'them ' : ''}before continuing.`,
+      sound: 'Frog',
+      title: '🚧 Lint Pilot 🚧',
+    })
+    return
+  }
+
+  notifier.notify({
+    message: 'All lint checks have passed. Your code is clean!',
+    sound: 'Purr',
+    title: '✅ Lint Pilot ✅',
+  })
+}
+
+export {
+  notifyResults,
+}
