+++ src/utils.mjs
@@ -1,0 +1,5 @@
+const pluralise = (message, count) => count === 1 ? message : `${message}s`
+
+export {
+  pluralise,
+}
